requirements.txt
streamlit
tensorflow==2.10.0
Pillow
numpy
matplotlib